# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Solar API",
    "description": "API interne pour la plateforme Solar (Association Le Trèfle 2.0). Supporte l'authentification par Bearer JWT ou API Key.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.beta.letrefle.org/",
      "description": "Serveur de production"
    },
    {
      "url": "http://localhost:3001",
      "description": "Serveur de développement local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "name": "token",
        "in": "header"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "name": "token",
        "in": "query"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "displayUsername": {
            "type": "string",
            "nullable": true
          },
          "newsletterSubscription": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "string",
            "description": "JSON string of permission array"
          },
          "weight": {
            "type": "integer"
          },
          "icon": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    },
    {
      "apiKeyQuery": []
    }
  ],
  "paths": {
...
    "/v1/admin/roles": {
      "get": {
        "summary": "Lister tous les rôles",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Liste des rôles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "roles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Role"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer ou mettre à jour un rôle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "permissions",
                  "weight"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "permissions": {
                    "type": "string"
                  },
                  "weight": {
                    "type": "integer"
                  },
                  "icon": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rôle enregistré"
          }
        }
      }
    },
    "/v1/admin/roles/{id}": {
      "delete": {
        "summary": "Supprimer un rôle",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rôle supprimé"
          }
        }
      }
    },
    "/v1/admin/users/{id}/role": {
      "post": {
        "summary": "Attribuer des rôles à un utilisateur",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roles"
                ],
                "properties": {
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rôles mis à jour"
          }
        }
      }
    },
...
    "/v1/auth/get-session": {
      "get": {
        "summary": "Récupérer la session utilisateur actuelle",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Session active",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.beta.letrefle.org/"
    }
  ]
}